% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FE_carriers}
\alias{FE_carriers}
\title{Information on Final Energy Carriers included in the M Matrix}
\format{
A sparse matrix of the size 222x9800. Each row is for a specific carrier for a specific accounting category (NTRA, NENE, etc.).
\describe{
\item{name}{Contains self-explanatory codes other than NENE (non-energy) & NTRA (non-transport)}
\item{unit}{All in TJ}
}
}
\source{
Extracted from EXIOBASE3 (provided by Arkaitz in Jan 2019)
}
\usage{
FE_carriers
}
\description{
A dataset containing names and units of energy carriers
}
\keyword{datasets}
