% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_exio_coicop_mapping.R
\name{load_exio_coicop_mapping}
\alias{load_exio_coicop_mapping}
\title{Load pre-made EXIO-COICOP mappings}
\usage{
load_exio_coicop_mapping()
}
\value{
map_cty list containing mappings for three countries ('IN', 'BR', 'ZA'), map_def for default mapping
}
\description{
Load pre-made EXIO-COICOP mappings
}
