% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_EXIOBASE3.R
\name{load_EXIOBASE3}
\alias{load_EXIOBASE3}
\title{Load EXIOBASE3 data}
\usage{
load_EXIOBASE3()
}
\value{
S, L, M, Y matrices as global variables
}
\description{
Load EXIOBASE3 data
}
\examples{
load_EXIOBASE3()
}
