% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{M}
\alias{M}
\title{Final Energy Intensity Matrix for EXIO FD vector}
\format{
A sparse matrix of the size 222x9800. Each row is for a specific carrier for a specific accounting category (NTRA, NENE, etc.).
}
\source{
Derived from EXIOBASE3 (provided by Arkaitz in Jan 2019)
}
\usage{
M
}
\description{
A dataset containing total final energy intensities \code{M=S⋅L}
diamonds.
}
\keyword{datasets}
