% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_to_exio.R
\name{map_to_exio}
\alias{map_to_exio}
\title{Convert COICOP consumption vector to EXIOBASE final demand vectors}
\usage{
map_to_exio(vec_coicop, country = "AT")
}
\arguments{
\item{vec_coicop}{A consumption vector in COICOP. It is a numerical vector with 164 entries.}

\item{country}{2-letter ISO country code}
}
\value{
A list with 'dom' and 'imp' entries, which are respectively for domestic and imported consumption vectors in EXIO product classification. Each vector has a length of 9800 (200 products x 49 countries).
}
\description{
Convert COICOP consumption vector to EXIOBASE final demand vectors
}
\examples{
final_cons_coicop = rep(1, 164)  # A dummy consumption vector for 164 categories
fd = map_to_exio(final_cons_coicop, 'IN')  # For India
footprint_dom = M \%*\% fd$dom
footprint_imp = M \%*\% fd$imp
}
