% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{M}
\alias{M}
\title{Final Energy Intensity Matrix for EXIO FD Vector}
\format{
A sparse matrix of the size 222x9800. Each row is for a specific carrier for a specific accounting category (NTRA, NENE, etc.).
}
\source{
Derived from EXIOBASE3 (provided by Arkaitz in Jan 2019)
}
\usage{
M
}
\description{
A dataset containing total final energy intensities \code{M=S⋅L}
}
\examples{
final_cons_coicop = rep(1, 164)  # 164 categories
fd_query = map_to_exio(final_cons_coicop, 'IN')
footprint_dom = M \%*\% fd_query$dom
}
\keyword{datasets}
